pyproject.toml
[project]
name = "vlamguard"
version = "1.0.0a1"
description = "Intelligent change risk engine for infrastructure changes"
readme = "README.md"
authors = [
    { name = "Elky Bachtiar", email = "bachtiare@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.10.0",
    "jsonschema>=4.23.0",
    "detect-secrets>=1.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
vlamguard = "vlamguard.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vlamguard"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]
